refactor(frontend): moved logic away from pages
