+++ new/src/frontend/acc2/src/router.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router";
-import { Home } from "./pages/home";
+import { HomePage } from "./pages/home";
 import { SetupPage } from "./pages/setup";
-import { Results } from "./pages/results";
+import { ResultsPage } from "./pages/results";
 import { Layout } from "./components/layout/layout";
 
 export const routes = ["setup", "results"] as const;
@@ -12,9 +12,9 @@ export const Router = () => {
     <BrowserRouter>
       <Routes>
         <Route element={<Layout />}>
-          <Route index element={<Home />} />
+          <Route index element={<HomePage />} />
           <Route path="setup" element={<SetupPage />} />
-          <Route path="results" element={<Results />} />
+          <Route path="results" element={<ResultsPage />} />
         </Route>
       </Routes>
     </BrowserRouter>
